+++ anvil_consortium_manager/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.12dev5"
+__version__ = "0.12dev6"