+++ anvil_consortium_manager/views.py
@@ -1,3 +1,9 @@
+import math
+
+import networkx as nx
+import numpy as np
+import plotly
+import plotly.graph_objects as go
 from dal import autocomplete
 from django import VERSION as DJANGO_VERSION
 from django.conf import settings
@@ -106,6 +112,176 @@ class AnVILAuditMixin:
             kwargs["audit_not_in_app"] = self.audit_results.get_not_in_app()
 
         return super().get_context_data(*args, **kwargs)
+
+
+class ManagedGroupGraphMixin:
+    """Mixin to add a plotly graph of group structure to context data."""
+
+    def get_graph(self):
+        """Return a graph of the group structure."""
+        raise NotImplementedError("You must override get_graph.")
+
+    def layout_graph(self):
+        """Lay out the nodes in the graph."""
+        # Networkx layout that requires graphviz:
+        # self.graph_layout = nx.drawing.nx_agraph.graphviz_layout(self.graph, prog="neato")
+        # Networkx layout that requires scipy:
+        # self.graph_layout = nx.kamada_kawai_layout(self.graph)
+        self.graph_layout = nx.spring_layout(self.graph)
+
+    def plot_graph(self):
+        """Create a plotly figure of the graph."""
+        point_size = 10
+
+        # Set up the figure.
+        layout = go.Layout(
+            height=700,
+            # showlegend=False,
+            xaxis=dict(showgrid=False, zeroline=False, showticklabels=False),
+            yaxis=dict(showgrid=False, zeroline=False, showticklabels=False),
+            legend=dict(
+                orientation="h",
+                yanchor="bottom",
+                y=1.02,
+                xanchor="right",
+                x=1,
+            ),
+        )
+
+        # Create the figure.
+        fig = go.Figure(layout=layout)
+
+        if self.graph:
+            # Group nodes as points.
+            node_x = []
+            node_y = []
+            node_labels = []
+            node_annotations = []
+            node_color = []
+            for node, d in self.graph.nodes(data=True):
+                x, y = self.graph_layout[node]
+                node_x.append(x)
+                node_y.append(y)
+                node_labels.append(
+                    node
+                    + "<br>Number of groups: {}<br>Number of accounts: {}".format(
+                        d["n_groups"], d["n_accounts"]
+                    )
+                )
+                node_annotations.append(
+                    go.layout.Annotation(
+                        dict(
+                            x=x,
+                            y=y,
+                            xref="x",
+                            yref="y",
+                            text=node,
+                            arrowhead=0,
+                            arrowcolor="#ccc",
+                        )
+                    )
+                )
+                node_color.append(math.log10(max(1, d["n_groups"] + d["n_accounts"])))
+
+            node_trace = go.Scatter(
+                x=node_x,
+                y=node_y,
+                mode="markers",
+                hoverinfo="text",
+                text=node_labels,
+                textposition="top center",
+                marker=dict(
+                    color=node_color,
+                    size=point_size,
+                    line_width=2,
+                    showscale=True,
+                    colorscale="YlGnBu",
+                    colorbar=dict(
+                        thickness=15,
+                        title="Group or account members",
+                        xanchor="left",
+                        titleside="right",
+                        tickmode="array",
+                        tickvals=[np.min(node_color), np.max(node_color)],
+                        ticktext=["Fewer", "More"],
+                        ticks="outside",
+                    ),
+                ),
+                name="managed groups",
+            )
+            fig.add_trace(node_trace)
+
+            # Group memberships as lines.
+            edge_x_member = []
+            edge_y_member = []
+            edge_x_admin = []
+            edge_y_admin = []
+            for u, v, e in self.graph.edges(data=True):
+                # Ignore direction.
+                x1, y1 = self.graph_layout[v]
+                x0, y0 = self.graph_layout[u]
+                if e["role"] == models.GroupGroupMembership.MEMBER:
+                    edge_x = edge_x_member
+                    edge_y = edge_y_member
+                elif e["role"] == models.GroupGroupMembership.ADMIN:
+                    edge_x = edge_x_admin
+                    edge_y = edge_y_admin
+                    # Reverse order so arrows go from child to parent instead of parent to child.
+                edge_x.append(x1)
+                edge_x.append(x0)
+                edge_x.append(None)
+                edge_y.append(y1)
+                edge_y.append(y0)
+                edge_y.append(None)
+
+            # Member relationships.
+            edge_trace_member = go.Scatter(
+                name="member role",
+                x=edge_x_member,
+                y=edge_y_member,
+                line=dict(width=0.5, color="#888"),
+                hoverinfo="none",
+                mode="lines+markers",
+                marker=dict(
+                    symbol="arrow",
+                    size=15,
+                    angleref="previous",
+                ),
+            )
+            fig.add_trace(edge_trace_member)
+
+            # Admin relationships.
+            edge_trace_admin = go.Scatter(
+                x=edge_x_admin,
+                y=edge_y_admin,
+                line=dict(width=2, color="#888"),
+                hoverinfo="none",
+                mode="lines+markers",
+                marker=dict(
+                    symbol="arrow",
+                    size=15,
+                    angleref="previous",
+                ),
+                name="admin role",
+            )
+            fig.add_trace(edge_trace_admin)
+
+            # Add group names as annotations.
+            fig.update_layout(
+                {"annotations": node_annotations},
+                margin=dict(l=20, r=20, t=50, b=20),
+                plot_bgcolor="#eee",
+            )
+
+        return fig
+
+    def get_context_data(self, **kwargs):
+        """Add the graph to the context data."""
+        context = super().get_context_data()
+        self.get_graph()
+        self.layout_graph()
+        context["graph"] = plotly.io.to_html(self.plot_graph(), full_html=False)
+        return context
 
 
 class Index(auth.AnVILConsortiumManagerViewRequired, TemplateView):
@@ -779,9 +955,29 @@ class AccountAudit(
         self.audit_results = models.Account.anvil_audit()
 
 
-class ManagedGroupDetail(auth.AnVILConsortiumManagerViewRequired, DetailView):
+class ManagedGroupDetail(
+    auth.AnVILConsortiumManagerViewRequired, ManagedGroupGraphMixin, DetailView
+):
     model = models.ManagedGroup
     slug_field = "name"
+
+    def get_graph(self):
+        self.graph = self.object.get_graph()
+
+    def plot_graph(self):
+        fig = super().plot_graph()
+        # Replot this group in a different color.
+        # Annotate this group.
+        fig.add_annotation(
+            x=self.graph_layout[self.object.name][0],
+            y=self.graph_layout[self.object.name][1],
+            text=self.object.name,
+            bgcolor="lightskyblue",
+            bordercolor="black",
+            showarrow=True,
+            arrowhead=1,
+        )
+        return fig
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -866,6 +1062,18 @@ class ManagedGroupList(auth.AnVILConsortiumManagerViewRequired, SingleTableView)
     model = models.ManagedGroup
     table_class = tables.ManagedGroupTable
     ordering = ("name",)
+
+
+class ManagedGroupVisualization(
+    auth.AnVILConsortiumManagerViewRequired, ManagedGroupGraphMixin, TemplateView
+):
+    """Display a visualization of all group relationships."""
+
+    template_name = "anvil_consortium_manager/managedgroup_visualization.html"
+
+    def get_graph(self):
+        G = models.ManagedGroup.get_full_graph()
+        self.graph = G
 
 
 class ManagedGroupDelete(